Move filter reducer into filter.js
