+++ desktop/apps/auction2/components/filter/FilterSort.js
@@ -2,7 +2,7 @@ import _ from 'underscore'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { connect } from 'react-redux'
-import { updateSort } from 'desktop/apps/auction2/actions'
+import { updateSort } from 'desktop/apps/auction2/actions/filter'
 
 function FilterSort (props) {
   const {
