+++ desktop/apps/auction2/components/filter/RangeSlider.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { Range } from 'rc-slider'
 import { connect } from 'react-redux'
 import { formatMoney } from 'accounting'
-import { updateEstimateRange, updateEstimateDisplay } from 'desktop/apps/auction2/actions'
+import { updateEstimateRange, updateEstimateDisplay } from 'desktop/apps/auction2/actions/filter'
 
 function RangeSlider (props) {
   const {
