+++ desktop/apps/auction2/reducers/app.js
@@ -1,0 +1,5 @@
+export const initialState = {}
+
+export default function appReducer (state = initialState, action = {}) {
+  return state
+}
